feat: move login check to login page
